refactor: separate scenario-specific e2e helpers from shared utilities
Move scenario-specific fake-gh factory functions and FakeDaemonHandle
out of tests/e2e/helpers/ and into xxx_fixtures.rs files co-located
with the test files that use them.

- helpers/env.rs shrinks from 577 to 245 lines (13 methods removed)
- FakeDaemonHandle moves to daemon/lifecycle_fixtures.rs
- Exposes setup_git_dirs / setup_empty_dirs as pub(crate) primitives
  so fixture modules can construct TestEnv without duplicating logic
- Cross-category methods (with_no_pr, with_pr_list, with_error,
  without_gh) remain in helpers/

Closes #291
