+++ tests/e2e/daemon/lifecycle.rs
@@ -5,7 +5,8 @@
 use assert_cmd::Command;
 use predicates::prelude::*;
 
-use super::super::helpers::{DaemonHandle, FakeDaemonHandle, TestEnv};
+use super::super::helpers::{DaemonHandle, TestEnv};
+use super::lifecycle_fixtures::FakeDaemonHandle;
 
 const BIN: &str = "merge-ready";
 const PROMPT_BIN: &str = "merge-ready-prompt";
