+++ tests/e2e/daemon/mod.rs
@@ -1,2 +1,3 @@
 mod lifecycle;
+mod lifecycle_fixtures;
 mod watch;
