+++ tests/e2e/scenarios/no_repo.rs
@@ -5,14 +5,15 @@
 use assert_cmd::Command;
 use predicates::prelude::*;
 
-use super::super::helpers::{DaemonHandle, TestEnv};
+use super::super::helpers::DaemonHandle;
+use super::no_repo_fixtures;
 
 const PROMPT_BIN: &str = "merge-ready-prompt";
 
 /// git リポジトリでない場合、何も出力しない
 #[test]
 fn test_no_git_remote_shows_nothing() {
-    let env = TestEnv::without_git_remote();
+    let env = no_repo_fixtures::without_git_remote();
     // daemon を起動して git リポジトリ外クエリを処理させる
     let _daemon = DaemonHandle::start(&env);
